refactor(ledger/tests): extract shared test utilities and address PR review feedback
- Create testUtils.ts with shared helpers, typed mutation wrappers, and
  getConvexErrorCode (now includes regex string fallback for all error codes)
- Update reservation.test.ts and concurrency.test.ts to import from testUtils
- Fix misleading "Redeem A 1,000" comments (actual amount is 2,000)
- Fix mintAndIssueMultiple JSDoc to reflect that validation is in mutations
